feat: se trasladan los componentes tabla y rechazar a shared #18 #19
