+++ src/app/pages/aprobacion-ordenespago/aprobacion-ordenespago-routing.module.ts
@@ -3,6 +3,7 @@ import { Routes, RouterModule } from '@angular/router';
 import { LayoutAprobacionordenComponent } from './components/layout-aprobacionorden/layout-aprobacionorden.component';
 import { BudgetApprobationComponent } from './components/budget-approbation/budget-approbation.component';
 import { AccountantApprobationComponent } from './components/accountant-approbation/accountant-approbation.component'
+import { OrderApprobationComponent } from './components/order-approbation/order-approbation.component';
 
 const routes: Routes = [
   {
@@ -10,12 +11,16 @@ const routes: Routes = [
     component: LayoutAprobacionordenComponent,
     children: [
       {
+        path: 'lista',
+        component: OrderApprobationComponent,
+      },
+      {
         path: 'contable',
         component: AccountantApprobationComponent,
       },
       {
         path: 'presupuestal',
-        component: BudgetApprobationComponent
+        component: BudgetApprobationComponent,
       }
     ],
   },
