+++ src/app/pages/aprobacion-ordenespago/aprobacion-ordenespago.module.ts
@@ -5,14 +5,19 @@ import { AprobacionOrdenespagoRoutingModule } from './aprobacion-ordenespago-rou
 import { LayoutAprobacionordenComponent } from './components/layout-aprobacionorden/layout-aprobacionorden.component';
 import { AccountantApprobationComponent } from './components/accountant-approbation/accountant-approbation.component';
 import { BudgetApprobationComponent } from './components/budget-approbation/budget-approbation.component';
+import { OrderApprobationComponent } from './components/order-approbation/order-approbation.component';
+import { SharedModule } from '../../shared/shared.module';
 
 
 @NgModule({
   declarations: [LayoutAprobacionordenComponent,
     AccountantApprobationComponent,
-    BudgetApprobationComponent],
+    BudgetApprobationComponent,
+    SharedModule,
+    OrderApprobationComponent],
   imports: [
     CommonModule,
+    SharedModule,
     AprobacionOrdenespagoRoutingModule
   ]
 })
