+++ src/app/pages/aprobacion-ordenespago/components/budget-approbation/budget-approbation.component.html
@@ -16,6 +16,11 @@
                     </div>
                     <div class="row">
                         <div class="col">
+                            <button (click)="regresar()">
+                                Volver al registro
+                            </button>
+                        </div>
+                        <div class="col">
                             <button type="button" class="btn btn-outline-dark px-3 py-2 my-2" (click)="rechazarAprobacion()">
                                 Rechazar
                             </button>
