+++ src/app/pages/aprobaciones/aprobaciones-routing.module.ts
@@ -15,7 +15,7 @@ const routes: Routes = [
       },
       {
         path: 'crear',
-        component: CreateApprobationComponent
+        component: CreateApprobationComponent,
       },
       {
         path: 'orden',
