+++ src/app/pages/aprobaciones/components/create-approbation/create-approbation.component.html
@@ -18,9 +18,6 @@
                         <div *ngIf=" nombreAprobacion === 'Relación de autorización'">
                             <ngx-relation-approbation></ngx-relation-approbation>
                         </div>
-                        <div *ngIf=" nombreAprobacion === 'Orden de pago'">
-                            <ngx-order-approbation></ngx-order-approbation>
-                        </div>
                         <div class="row">
                             <button mat-button matStepperPrevious type="button" class="btn btn-primary px-3 py-2 my-2 text-light">
                                 Volver al registro
