+++ src/app/pages/aprobaciones/interfaces/interfaces.ts
@@ -201,41 +201,3 @@ export const DATOS_CREACION: any = [
             estado: 'activo'
          }
 ];
-export const DATOS_ORDENPAGO: any = [
-    {
-        vigencia: '2019',
-        consecutivo: '21783',
-        tipoID  : 'CC',
-        numeroID: '120398',
-        nombre: 'Ester'
-    },
-    {
-        vigencia: '2021',
-        consecutivo: '2453',
-        tipoID  : 'CC',
-        numeroID: '87324',
-        nombre: 'Juan'
-     },
-     {
-        vigencia: '2020',
-        consecutivo: '21783',
-        tipoID  : 'CC',
-        numeroID: '379989',
-        nombre: 'Lina'
-     },
-     {
-        vigencia: '2019',
-        consecutivo: '21233',
-        tipoID  : 'CC',
-        numeroID: '5465598',
-        nombre: 'Ana'
-     },
-     {
-        vigencia: '2019',
-        consecutivo: '21908',
-        tipoID  : 'CC',
-        numeroID: '120455',
-        nombre: 'Nancy'
-     },
-];
-
