+++ src/app/shared/components/rechazar/rechazar.component.scss
