+++ src/app/shared/components/table/table.component.html
